Move migration types to Core
