+++ src/Database/Persist/Migration/Operation.hs
@@ -12,13 +12,7 @@ Defines the Operation data types.
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
 module Database.Persist.Migration.Operation
-  ( Version
-  , OperationPath
-  , (~>)
-  , Migration
-  , MigrationPath(..)
-  , opPath
-  , Operation(..)
+  ( Operation(..)
   , validateOperation
   ) where
 
@@ -30,34 +24,6 @@ import Database.Persist.Migration.Operation.Types
 import Database.Persist.Migration.Utils.Data (isConstr)
 import Database.Persist.Migration.Utils.Sql (MigrateSql)
 import Database.Persist.Sql (PersistValue, SqlPersistT)
-
--- | The version of a database. An operation migrates from the given version to another version.
---
--- The version must be increasing, such that the lowest version is the first version and the highest
--- version is the most up-to-date version.
---
--- A version represents a version of the database schema. In other words, any set of operations
--- taken to get to version X *MUST* all result in the same database schema.
-type Version = Int
-
--- | The path that an operation takes.
-type OperationPath = (Version, Version)
-
--- | An infix constructor for 'OperationPath'.
-(~>) :: Version -> Version -> OperationPath
-(~>) = (,)
-
--- | A migration list that defines operations to manually migrate a database schema.
-type Migration = [MigrationPath]
-
--- | A path representing the operations needed to run to get from one version of the database schema
--- to the next.
-data MigrationPath = OperationPath := [Operation]
-  deriving (Show)
-
--- | Get the OperationPath in the MigrationPath.
-opPath :: MigrationPath -> OperationPath
-opPath (path := _) = path
 
 -- | An operation that can be migrated.
 data Operation
